initial file
Created initial file.  It consists of a nav menu, jumbotron, table with headers only and an input form.  Javascript to follow.  Javascript should make a connection to Firebase to record and retrieve the object data.  It would be nice to have this hook into NJ Transit or something like that to pull in real train data.